Added uniform rng in [0,1), scalar product and microcanonical update
Added the functions in title, with tests for each of them. Also, the work has been organized in 3 main folders: functions_test, include and lib. In the first one all the single functions tests will be performed, in the second one all the c files useful for the final simulation will be inserted (without main to test them, which will be in the first one) and in the last one all the headers will be inserted.