Added function to compute magnetization vector
The function for computing magnetization vector has been implemented and tested in magnetization.c. Such test consists of a sum of all equal vector of the cubic lattice, so it's very simple. Also, it's NOT in functions.c and in function.h, to avoid problems with the merging of the two branches. After merging the branches, it has to be tested in the usual way (test in function_tests and function defined in the lib and in the include functions files).