Added function to compute total energy of the system
The function for computing E has been implemented and tested in energy.c. It's NOT in functions.c and in function.h, to avoid problems with the merging of the two branches. After merging the branches, it has to be tested in the usual way (test in function_tests and function defined in the lib and in the include functions files).