Starting the construction of the main for 3D O(2) system
A main that take an input file and an output file (this one is optional, if not given stdout is choosen as default place where to print output) has been implemented. Such main scans the input file, looking for the needed parameters, and prints in output the number found for each or abort the simulation if one of the value is not found in the given input file. Also, the main allocate and randomly inizializes the lattice, giving an ok when everything went fine.