Added the option to choose the formatting of the .dat
A keyword to choose the formatting of the data.dat file has been added. Now both the o2_metro*.c files look for the keywork 'output_data_format' in the input file and, as usual, if it's not given or a incompatible one is given, the simulation is aborted. The compatible keywords are two: 'minimal' formatting leads to a file with just mx, my and energy per site, while 'complete' formatting leads to the usual output .dat, with step, lattice indexes, s_old, s_new, mx, my, energy and type of step (microcanonical or metropolis). Minimal-formatted files have size ~1/3 of the complete-formatted ones. For what concernes the run.sh file, it has been changed to be compatible with this new way of writing input.