Now it's possible to write .bin files instead of .dat
The 'minimal' keyword is now used to ask for the writing of the associated 'minimal' data output (mx my E_per_site) in a binary file, instead of the usual txt. In this way the space occupied by data is reduced of ~1/2. The 'complete' option is still available to get all the data in a readeable file format, .dat. Maybe all of this should be implemented in a function, to make code more compact and readeable...