Added the possibility to choose for RNG seed
Now it's possible, thanks to the keyword 'seed', to choose for a custom seed or for a seed dependent on time.h: if the word after 'seed' is 'time', than the function time(NULL) is used to get a seed, otherwise what's after 'seed' is converted from string to a long unsigned integer, using atoi(). By using such new keyword, i.e. by setting a fixed seed, it has been possible to check if the binary formatted output (minimal) was having the same values for mx, my and E as txt output files (complete).