Now blocking.py is runnable from terminal
blocking.py is runnable via terminal, taking as input --datafile and --plot_name. The first argument is the .dat file to analize with the blocking procedure, while the second one is the plot file name. Right now all of this it has been implemented just for m_y. Next steps are implement it for m_x and energy and create a .sh to perform a for cycle over all the files.