Modified blocking.py
An intermediate version of the script has been implemented. Now it's possible to add a new argument from command line and all the useful variables for our case of interest (mx, my, E, modules) are correctly allocated, computed (for the modules) and checked (again, modules). Some work is still needed.