Implemented core of Jackknife procedure for Binder and Variance
The core of the Jackknife procedure has been implemented. A test has been performed by computing the variance of the sample variance of a gaussian sample. We can perform such test because we know that the distribution of the sample variance of a gaussian dataset is linked to a chi squared distribution, so we can analitically compute the the variance of the sample variance.