Added function perform_jackknife_blocking
The function perform_jackknife_blocking takes an input path, performs blocking and apply jackknife for the various blocked data obtained. After that, it saves all the variances obtained for each csv file processed in a new csv, together with a column containing the blocksize for the variance, L and beta (these last two columns are redundant).