Added perform_jackknife_blocking to perform jk+blocking for a single blocksize
The function perform_jackknife_blocking is very similar to the perform_jackknife_blocking_analysis, but without parallelization and working with just a single blocksize. The config file .yaml has been modified coherently to be clearer and to include also the parameters of the new added function. A similar thing has been done with the interface that asks to the users what parameters they want to insert, and with the main in data_analysis.